+++ src/commands.js
@@ -1,0 +1,11 @@
+import { showCommand } from "./show.js";
+import { smashCommand } from "./smash.js";
+import { scanCommand } from "./scan.js";
+import { readCommand } from "./read.js";
+
+export {
+	showCommand as show,
+	smashCommand as smash,
+	scanCommand as scan,
+	readCommand as read,
+};